# various performance and memory optimization