--- a/Project.toml
+++ b/Project.toml
@@ -7,10 +7,12 @@ version = "0.2.5"
 ImageCore = "a09fc81d-aa75-5fe9-8630-4744c3626534"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+TiledIteration = "06e1c1a7-607b-532d-9fad-de7d9aa2abac"
 
 [compat]
 ImageCore = "0.8.1, 0.9"
 OffsetArrays = "1"
+TiledIteration = "0.3"
 julia = "1"
 
 [extras]