The result visually looks "correct" to me so a double-check would be the best, @adrhill it would be better if you can write some correctness test in the future.

```julia
using DitherPunk
using TestImages, ImageTransformations, ImageShow, ImageCore

img = imresize(testimage("cameraman"), (240, 240))

@btime bayer_dithering($img);
# master: 1.490 ms (15 allocations: 969.02 KiB)
# PR: 187.368 μs (15 allocations: 233.62 KiB)

@btime bayer_dithering($img, to_linear=true);
# master: 3.190 ms (15 allocations: 969.02 KiB)
# PR: 279.525 μs (15 allocations: 233.62 KiB)


@btime simple_error_diffusion($img);
# master: 743.018 μs (4 allocations: 232.31 KiB)
# PR: 702.736 μs (5 allocations: 232.41 KiB)

@btime simple_error_diffusion($img, to_linear=true);
# master: 1.753 ms (4 allocations: 232.31 KiB)
# PR: 766.562 μs (5 allocations: 232.41 KiB)


@btime threshold_dithering($img, to_linear=true);
# master: 1.784 ms (7 allocations: 180.39 KiB)
# PR: 155.320 μs (19 allocations: 12.34 KiB)
```