@johnnychen94 2021-05-06
The `@inbounds` annotation in ff21dd7 removes the boundary check during `getindex` operations (e.g., `img[p]`)

```julia
@btime simple_error_diffusion($img);
# master: 743.018 μs (4 allocations: 232.31 KiB)
# before: 702.736 μs (5 allocations: 232.41 KiB)
# after: 481.427 μs (5 allocations: 232.41 KiB)

@btime simple_error_diffusion($img, to_linear=true);
# master: 1.753 ms (4 allocations: 232.31 KiB)
# before: 766.562 μs (5 allocations: 232.41 KiB)
# after: 562.010 μs (5 allocations: 232.41 KiB)
```