@adrhill 2021-05-06
> Some packages in JuliaImages use `ReferenceTests` as a tool to write regression test. For example, [ImageBinarization.jl](https://github.com/zygmuntszpak/ImageBinarization.jl/blob/afd449f932cba45495d2eb6099d0202e09c5f039/test/algorithms/adaptive_threshold.jl#L53-L56)

Thanks a lot, that's good to know! I'll open an issue.